refactor(ai-diagnostic): extract pure analytics from predictive_maintenance
PredictiveMaintenanceService mixed pure numeric logic (trend/regression, fan
failure probability + status, prediction confidence, recommendation text) into
the service and had no coverage over it. Extracted, test-first:

  service/maintenance_models.py    - MetricDataPoint, TrendAnalysis,
                                     FailurePrediction, ComponentHealth
  service/maintenance_analytics.py - analyze_trend + assess/calculate/confidence
                                     + the fan/temperature/battery/fiber
                                     recommendation text (pure functions)

The service methods are now thin delegators (signatures unchanged, so every
call site and the singleton stay put); dataclasses are re-exported for existing
importers. Added 14 characterisation tests over the analytics (regression
direction/slope/r-squared, failure probability + TTF, health thresholds,
confidence tiers). Dropped now-dead imports (numpy, statistics, dataclass,
field). 912 -> 655 LOC; magic CV/slope cutoffs named. All 201 ai-diagnostic
tests pass.
